database.yml: add pkg to the list
Even though it is an essential component of FreeBSD - for ports and for
the base system since the introduction of pkgbase - pkg itself is
currently developed like a third-party tool. Its development is hosted
outside of FreeBSD's base repository.

The directory listed here is that of the placeholder implemented in
FreeBSD's base system, for the import of the real binary for pkg.
pkg instantly jumps at the top of the chain for security, as it is
relevant for the build system, for the integrity of the whole system,
and functions as a network client.

As suggested by cperciva@; thank you!