src/versions: simplify the reporting
This ensures a consistent behaviour when unable to collect a version
number. Not finding a version number is now considered a valid output.

The goal is to pass the tests even when some versions could not be
obtained. Beyond the GitHub workflow file (using 14.3 at the moment)
this should allow for more flexibility when collecting information on
past and future releases, or when disabling or adding some components.